refactor: dvd content and contents.
